Add files via upload
A function scrabble(tiles, word) that takes a set of scrabble tiles and a word and returns True if the word can be formed from some / all of the tiles and False otherwise. The tiles are standard scrabble tiles, A-Z(There will be a total of 26 tiles one for each letter) and blank tiles(2 in number). A blank tile can be considered as any one letter for forming the word. The program should not be case-sensitive because scrabble tiles are not. The blank tile should be represented as '?' in the input. The number of tiles drawn at a time (input to the program) is at least one.